Specify minimum number of zeros for SDF file format
The recent changes for looping through all possible numbers of
zeros for the SDF file representation introduced a bug.
If the number of digits used in the file format is less than the
number of digits required to represent the requested file number
then the format becomes '*.sdf' and the first match file will be
returned, rather than the snapshot number requested.
This change fixes that issue.

It fixes Lare3d CCPForge bug #1450 and also fixes issue #2 on
the Gitlab SDF issue tracker.